Edit some field names and behavior of models
I edited some variable names (for example: "datetime_created" to "created") because information in these fields was a bit unnecessary as
I see now. Then I edited serializers of orders to make them more flexible in work with related fields. Afterwards, I created the .create() method
for serializers. There's a real need in redefining it, because this is the only way in Django Rest Framework and Django to create needed
kind of relationship between models.