Add API for buying products (#808d5ffb)
This commit changes a lot in the behavior of the project. The main goal of this commit is to imitate work of the real payment service. I made a few
assumptions to achieve this, for example I even started sending secret key with the payment data :) It has a sense just because that is the
only way to imitate the confirming of authenticity of the request. Secondly, I even added FakePaymentService view to achieve this goal. This
decision of mine should separate where the part of the payment service begins and where the business logic should run. I also added a new
permission that works only with payments. There is a bunch of the small changes in project with this commit, but they are not that important
to describe them apart.