Restore the previous behavior of OrderSerializer.create()
I found that requesting the database in the loop is needed behavior since we must ensure that every product that an user asks is existing.
Previous approach is just filtering the IDs and does not throw any error if an user passed a product that is not the database, so I've
decided to return the database requests in the loop.