Edit RegisterSerializer to validate the uniqueness
I slowly add the token authorization to the project and because the DRF implementation of it is strongly related with Django User model, it
throws ugly exceptions when user provide an not unique username and email. With DRF serializer validation mechanism these exceptions
become more human-readable, so I put the validators here.