Add the token auth to the users app (#1603b2ba)
I finally added the token authorization to one of the apps. Now the users can access the endpoints that request authorization by providing
their login and pass or using their tokens.

I also added BrowsableApiRenderer into an every view, so now the API could be discoverable even from browser.

The last thing is I changed the tests a little, so now they work with the token authorization instead of the basic auth.