Add token authorization to tests (#1603b2ba)
Because of the reason that I do not fully understand, the adding of a token authorization
is somehow broke the standart approach of an authorization in tests, which is used to be
ClientApi.login() method. Now client requires to pass the credentials explicitely.
I achieve this by using the approach where we provide credentials on every request to the server.

This approach basically looks like this:
client.get('/some_endpoint_that_use_authorization', **test_data, HTTP_AUTHORIZATION = 'Token' + http_token)

As I made sure further, doing the manual API testing with curl, the standart authorization that use
login:password is still alive and works, but obstinate APITestCase.client.login() method does not
trust in this. This is why I add that ugly string to every request a test do.