Fix bug with permission denied
It was comparing the request.user.id value with the value of order.customer.id, but they're represented as different types in MySQL.
Because of that, it was always returning False and the permission was comparing them.
That's why it now compares only their values (using "==" operator) and not their types (using "is" keyword).