Fix payments tests
Not so long ago, the project switched from the SQLite to the MySQL database with the InnoDB engine.
It has several drawbacks with working with object IDs inside the database, so I had to rebuild the tests so that they work with any ID given by DB.
More than that, it was still using OrderStatus fields, but it was deleted recently.